Use the minimum time for ordering commit-sets and the commit time to workaround the problem shown below.
When commit A, B, C happen in short time (shorter than FUZZ_SEC) and the
author and the commit log of A and C is same, the cvs2{git,svn}dump script
put A and B into one commit-set and treats them in order B, (A, C).  If B
is related to A, this order causes problem.

Typically this probelm likely happens like bellow pattern:

  A: commit by mistake
  B: cancel 1
  C: commit correctly

Putting A and C into a commit-set is not avoidable.  But treating them
in order "(A, C), B" is better to mitigate the problem.